requirements.txt
aniso8601==9.0.1
apispec==5.2.1
apispec-webframeworks==0.5.2
attrs==21.4.0
click==8.1.3
Flask==2.1.2
Flask-PyMongo==2.3.0
flask-restplus==0.13.0
gunicorn==20.1.0
importlib-metadata==4.11.3
importlib-resources==5.7.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.5.1
MarkupSafe==2.1.1
marshmallow==3.15.0
packaging==21.3
passlib==1.7.4
PyJWT==2.4.0
pymongo==4.1.1
pyparsing==3.0.9
pyrsistent==0.18.1
pytz==2022.1
PyYAML==6.0
six==1.16.0
Werkzeug==2.1.2
zipp==3.8.0
